requirements.txt
awesome-slugify==1.6.5
captcha==0.2.1
click==6.6
defusedxml==0.4.1
dj-database-url==0.4.0
Django==1.9.7
django-bootstrap-form==3.2.1
django-braces==1.10.0
django-embed-video==1.1.0
django-haystack==2.4.1
django-ipware==1.1.3
django-markdown-deux==1.0.5
django-markdownx==1.3
django-memcache-status==1.2
django-nested-admin==3.0.12
django-nested-formset==0.1.4
django-simple-captcha==0.5.3
django-webpack-loader==0.3.3
django-widget-tweaks==1.4.1
djangorestframework==3.5.4
djorm-ext-pgfulltext==0.10
elasticsearch==2.2.0
Flask==0.11.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.6
markdown2==2.3.0
MarkupSafe==0.23
oauthlib==1.0.3
pendulum==0.6.2
Pillow==3.3.0
psycopg2==2.6.2
PyJWT==1.4.0
python-dateutil==2.5.3
python-memcached==1.58
python-monkey-business==1.0.0
python-slugify==1.1.4
python-social-auth==0.2.12
python3-openid==3.0.9
pytz==2016.7
regex==2017.1.12
requests==2.9.1
requests-oauthlib==0.6.1
rest-framework-generic-relations==1.1.0
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.2.1
sorl-thumbnail==12.3
tzlocal==1.2.2
Unidecode==0.4.18
urllib3==1.14
Werkzeug==0.11.10
whitenoise==2.0.6
Whoosh==2.7.2
